refactor: アイデアメモ詳細・編集画面Container/Presentationalパターン廃止 #175
