+++ front/src/app/idea-memos/[uuid]/IdeaMemosUuid.module.scss
